Rydder litt i redirect responses
